Render `script` and `shell` as literals, fix workspace-cleanup bug
To make the yaml more readable when it's using long
shell scripts, they should be rendered as literals.

This works for the majority of shell scripts, although
the yaml dump refuses to render some strings as literals
when they look like they should be fine. This isn't a bug
with wrecker, the data is going to the yaml dumper correctly,
but there's something in the string that the dumper doesn't
like

Also, fixed a small bug in the workspace-cleanup conversion.
The clean-if field should have been a list of dicts, not a dict.